+++ packages/utils/src/remove-duplicates.ts
@@ -3,17 +3,18 @@ import prettier from 'prettier'
 import { ITokenMetadata } from './types'
 import { METADATA_PATH } from './constants'
 
+const geckoCoins: ITokenMetadata[] = JSON.parse(
+  fs.readFileSync(METADATA_PATH, 'utf-8'),
+)
+
 /**
- * removes duplicates from the tokens.json file. 
- * finds the duplicated by symbol 
+ * removes duplicates from the tokens.json file.
+ * finds the duplicated by symbol
  * removes the object with the lowest marketCapRank
  * this assumes if two icons share the same symbol, they should have the same icon
  * also assumes that the marketCapRank is a good data to use to determine duplicates
  */
 export const removeDuplicates = async () => {
-  const geckoCoins: ITokenMetadata[] = JSON.parse(
-    fs.readFileSync(METADATA_PATH, 'utf-8'),
-  )
   const removedIds: string[] = []
 
   const processedCoins = geckoCoins.reduce(
@@ -39,7 +40,21 @@ export const removeDuplicates = async () => {
   )
 
   // Convert the processed coins back to an array
-  const uniqueCoins = Object.values(processedCoins)
+  // clean the addresses
+  const uniqueCoins = Object.values(processedCoins).map((item) => {
+    const cleanedAddresses = Object.entries(item.addresses || {})
+      .filter(([key, value]) => value)
+      .reduce((acc, [key, value]) => ({ ...acc, [key]: value }), {})
+
+    return { ...item, addresses: cleanedAddresses }
+  })
+
+  // .map((item) => {
+  //   const cleanedAddresses = Object.fromEntries(
+  //     Object.entries(item.addresses || {}).filter(([key, _]) => key !== ''),
+  //   )
+  //   return { ...item, addresses: cleanedAddresses }
+  // })
 
   // Write the modified data back to the file
   fs.writeFileSync(
@@ -50,10 +65,7 @@ export const removeDuplicates = async () => {
     'utf-8',
   )
 
-  console.log(
-    `Removed ${removedIds.length} duplicates from tokens.json`,
-  )
+  console.log(`Removed ${removedIds.length} duplicates from tokens.json`)
 }
-
 
 removeDuplicates().catch(console.error)