+++ packages/utils/src/scripts/update-metadata.ts
@@ -1,71 +1,101 @@
-import fs from 'fs';
-import path from 'path';
-import prettier from 'prettier';
-import { INetworkMetadata, ITokenMetadata } from '../types';
-import getCoinByID from './gecko/get-coin-by-id';
+import fs from 'fs'
+import path from 'path'
+import prettier from 'prettier'
+import { INetworkMetadata, ITokenMetadata } from '../types'
+import getCoinByID from './gecko/get-coin-by-id'
 
-const appendToNetworksJson = async (network: INetworkMetadata): Promise<void> => {
-  const jsonPath = path.resolve(process.cwd(), 'packages/core/src/metadata/networks.json');
+const appendToNetworksJson = async (
+  network: INetworkMetadata,
+): Promise<void> => {
+  const jsonPath = path.resolve(
+    process.cwd(),
+    'packages/core/src/metadata/networks.json',
+  )
 
-  let existingMetadata: INetworkMetadata[] = [];
+  let existingMetadata: INetworkMetadata[] = []
   if (fs.existsSync(jsonPath)) {
-    const fileContent = fs.readFileSync(jsonPath, 'utf-8');
-    existingMetadata = JSON.parse(fileContent) as INetworkMetadata[];
+    const fileContent = fs.readFileSync(jsonPath, 'utf-8')
+    existingMetadata = JSON.parse(fileContent) as INetworkMetadata[]
   }
 
-  const existingNetworkIndex = existingMetadata.findIndex((n) => n.id === network.id);
+  const existingNetworkIndex = existingMetadata.findIndex(
+    (n) => n.id === network.id,
+  )
 
   if (existingNetworkIndex > -1) {
-    const currentNetwork = existingMetadata[existingNetworkIndex];
+    const currentNetwork = existingMetadata[existingNetworkIndex]
     if (currentNetwork && currentNetwork.variants) {
       network.variants.forEach((variant) => {
         if (!currentNetwork.variants.includes(variant)) {
-          currentNetwork.variants.push(variant);
+          currentNetwork.variants.push(variant)
         }
-      });
+      })
     }
   } else {
-    existingMetadata.push(network);
+    existingMetadata.push(network)
   }
 
-  const formatted = await prettier.format(JSON.stringify(existingMetadata), { parser: 'json' });
-  fs.writeFileSync(jsonPath, formatted);
-};
+  const formatted = await prettier.format(JSON.stringify(existingMetadata), {
+    parser: 'json',
+  })
+  fs.writeFileSync(jsonPath, formatted)
+}
 
 const appendToTokensJson = async (coin: ITokenMetadata): Promise<void> => {
-  const jsonPath = path.resolve(process.cwd(), 'packages/core/src/metadata/tokens.json')
-  let existingMetadata: ITokenMetadata[] = [];
+  const jsonPath = path.resolve(
+    process.cwd(),
+    'packages/core/src/metadata/tokens.json',
+  )
+  let existingMetadata: ITokenMetadata[] = []
   if (fs.existsSync(jsonPath)) {
-    const fileContent = fs.readFileSync(jsonPath, 'utf-8');
-    existingMetadata = JSON.parse(fileContent) as ITokenMetadata[];
+    const fileContent = fs.readFileSync(jsonPath, 'utf-8')
+    existingMetadata = JSON.parse(fileContent) as ITokenMetadata[]
   }
 
-  const existingCoinIndex = existingMetadata.findIndex((c) => c.id === coin.id);
+  const existingCoinIndex = existingMetadata.findIndex((c) => c.id === coin.id)
 
   if (existingCoinIndex > -1) {
-    const currentCoin = existingMetadata[existingCoinIndex];
+    const currentCoin = existingMetadata[existingCoinIndex]
     if (currentCoin && currentCoin.variants) {
       coin.variants.forEach((variant) => {
         if (!currentCoin.variants.includes(variant)) {
-          currentCoin.variants.push(variant);
+          currentCoin.variants.push(variant)
         }
-      });
+      })
     }
   } else {
-    existingMetadata.push(coin);
+    existingMetadata.push(coin)
   }
 
-  fs.writeFileSync(jsonPath, await prettier.format(JSON.stringify(existingMetadata), { parser: 'json' }));
-};
+  fs.writeFileSync(
+    jsonPath,
+    await prettier.format(JSON.stringify(existingMetadata), { parser: 'json' }),
+  )
+}
 
-const processSVGFile = async (file: string, type: 'token' | 'network'): Promise<void> => {
-  const fileName = type === 'token' ? path.basename(file, '.svg').toUpperCase() : path.basename(file, '.svg');
+const processSVGFile = async (
+  file: string,
+  type: 'token' | 'network',
+): Promise<void> => {
+  const fileName =
+    type === 'token'
+      ? path.basename(file, '.svg').toUpperCase()
+      : path.basename(file, '.svg')
   console.log(fileName)
-  const geckoCoins = JSON.parse(fs.readFileSync(path.join(__dirname, './gecko/gecko-coins.json'), 'utf8'));
-  const geckoNetworks = JSON.parse(fs.readFileSync(path.join(__dirname, './gecko/gecko-networks.json'), 'utf8'));
+  const geckoCoins = JSON.parse(
+    fs.readFileSync(path.join(__dirname, './gecko/gecko-coins.json'), 'utf8'),
+  )
+  const geckoNetworks = JSON.parse(
+    fs.readFileSync(
+      path.join(__dirname, './gecko/gecko-networks.json'),
+      'utf8',
+    ),
+  )
 
   if (type === 'token') {
-    const foundCoin = geckoCoins.find((coin: any) => coin.symbol.toUpperCase() === fileName);
+    const foundCoin = geckoCoins.find(
+      (coin: any) => coin.symbol.toUpperCase() === fileName,
+    )
     if (foundCoin) {
       const tokenIcon: ITokenMetadata = {
         id: foundCoin.id,
@@ -74,21 +104,23 @@ const processSVGFile = async (file: string, type: 'token' | 'network'): Promise<
         variants: file.includes('/branded/') ? ['branded'] : ['mono'],
         marketCapRank: 0, // Will be fetched below
         addresses: {}, // Will be fetched below
-      };
+      }
       console.log('hey')
       console.log(file)
-      const data = await getCoinByID(foundCoin.id);
-      tokenIcon.addresses = data.platforms;
-      tokenIcon.marketCapRank = data.market_cap_rank;
-      console.log('tokenIcon', tokenIcon);
+      const data = await getCoinByID(foundCoin.id)
+      tokenIcon.addresses = data.platforms
+      tokenIcon.marketCapRank = data.market_cap_rank
+      console.log('tokenIcon', tokenIcon)
 
-      await appendToTokensJson(tokenIcon);
+      await appendToTokensJson(tokenIcon)
     }
   } else if (type === 'network') {
-    const foundNetwork = geckoNetworks.find((network: any) => 
-      network.id.toUpperCase() === fileName ||
-      network.shortname.toUpperCase() === fileName ||
-      network.name.toUpperCase() === fileName);
+    const foundNetwork = geckoNetworks.find(
+      (network: any) =>
+        network.id.toUpperCase() === fileName ||
+        network.shortname.toUpperCase() === fileName ||
+        network.name.toUpperCase() === fileName,
+    )
 
     if (foundNetwork) {
       const networkIcon: INetworkMetadata = {
@@ -97,26 +129,27 @@ const processSVGFile = async (file: string, type: 'token' | 'network'): Promise<
         shortname: foundNetwork.shortname,
         variants: file.includes('/branded/') ? ['branded'] : ['mono'],
         nativeCoinId: foundNetwork.native_coin_id,
-      };
+      }
       console.log('networkIcon', networkIcon)
 
-      await appendToNetworksJson(networkIcon);
+      await appendToNetworksJson(networkIcon)
     }
   }
-};
+}
 
 const main = async (): Promise<void> => {
-  const newSVGFiles: { file: string; type: 'token' | 'network' }[] = process.argv.slice(2).map(file => ({
-    file, 
-    type: file.includes('/tokens/') ? 'token' : 'network',
-  }));
-  
-  for (const { file, type } of newSVGFiles) {
-    await processSVGFile(file, type);
-  }
-};
+  const newSVGFiles: { file: string; type: 'token' | 'network' }[] =
+    process.argv.slice(2).map((file) => ({
+      file,
+      type: file.includes('/tokens/') ? 'token' : 'network',
+    }))
 
-main().catch(console.error);
+  for (const { file, type } of newSVGFiles) {
+    await processSVGFile(file, type)
+  }
+}
+
+main().catch(console.error)
 
 // use this on development
 // function getNewSVGFiles(): { file: string; type: string }[] {