+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @token-icons/core
 
+## 1.8.1
+
+### Patch Changes
+
+- fix ITokenMetadata types
+
 ## 1.8.0
 
 ### Minor Changes